Rename Untitled5.ipynb to UsingGeoPy.ipynb
This is a little exercise in which I used Geopy to create an app that searches latitudes and longitudes to locate places. 

Thus, the app will ask you for a place (a city in Colombia... Pasto, Cali, Cartagena, Medellin, or any other) and if the place is located in the database it will send you a response with the lat, and longitude. Pretty useful to locate places and feed a map.