Rename Untitled2.ipynb to PythonToSQL.ipynb
This is an example of how to use Python to connect with a server and to use Geopy and mysql.connector to populate a table. 

The application will ask you to fill out some spaces like the name of the client, address, longitude, and latitude where the client is located. 

If the connection is done properly python will fill out the table and close the connection again.